backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
openai==1.70.0
langchain==0.3.23
faiss-cpu==1.7.4
pydantic==2.11.2 